Fixed DesignTimeData on Busget from int to long.

+++ DungeonMasterVault.Services/DataServices/SampleDataService.cs
@@ -61,7 +61,7 @@
                     ID = "R" + i,
                     Name = "R" + i + ". Sample Encounter",
                     Adventure = "Runtime Adventure",
-                    Budget = new int?(100)
+                    Budget = new long?(100)
                 };
                 encounters.Add(encounter);
             }
